+++ pages/index.js
@@ -1,4 +1,4 @@
-import Article from "@/components/Article.js/ArticleList";
+import ArticleList from "@/components/ArticleList/ArticleList";
 import Navigation from "@/components/Navigation/Navigation";
 
 const fetcher = (url) => fetch(url).then((res) => res.json());
@@ -7,7 +7,7 @@ export default function Home() {
   return (
     <>
       <Navigation />
-      <Article />
+      <ArticleList />
     </>
   );
 }
